requirements.txt
streamlit
textblob
opencv-python
pillow
gTTS